building/02/19-2

+++ uni-place-backend/src/main/java/org/myweb/uniplace/domain/property/domain/entity/Building.java
@@ -23,7 +23,7 @@ public class Building extends ActivateEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long buildingId;
+    private Integer buildingId;
 
     @Column(nullable = false, unique = true)
     private String buildingNm;
